fix(map): _lastClusterZoomTier int?, не double? — int не assignable в double
CI analyze: invalid_assignment в map_screen.dart:472

+++ lib/features/map_screen/map_screen.dart
@@ -460,7 +460,7 @@ class _MapScreenState extends State<MapScreen> {
     }
   }
 
-  double? _lastClusterZoomTier;
+  int? _lastClusterZoomTier;
 
   void _onMapPosition(MapCamera camera, bool hasGesture) {
     // Дебаунс: грузим слои после остановки камеры.
